重构 qma config 命令

+++ lib/qma/config.rb
@@ -1,23 +1,22 @@
 require 'yaml'
 
-
 module QMA
   class Config
     attr_reader :path, :data
 
     def initialize(path = nil)
-      if path
+      if path.to_s.empty?
+        load_default_config
+      else
         raise QMA::NotFoundError, path unless File.exist?(path)
         load(path)
-      else
-        load_default_config
       end
 
       migratate_old_data if old_data?
     end
 
     def key
-    @data.try(:[], 'key')
+      @data.try(:[], 'key')
     end
 
     def key=(key)
@@ -37,8 +36,8 @@ module QMA
     end
 
     def hosts=(host)
-      external_host = host
-      intranet_host = host
+      self.external_host = host
+      self.intranet_host = host
     end
 
     def external_host=(host)
@@ -56,7 +55,7 @@ module QMA
     end
 
     def load_default_config
-      FileUtils.cp template_config_file, default_path
+      FileUtils.cp template_config_file, default_path unless File.exist?(default_path)
       load(default_path)
     end
 
@@ -64,6 +63,8 @@ module QMA
       File.open(@path, 'w') do |f|
         f.write @data.to_yaml
       end
+
+      self
     end
 
     def save!
@@ -76,48 +77,46 @@ module QMA
 
     private
 
-      def template_config_file
-        source_path = File.expand_path('../../../config', __FILE__)
-        source_file = File.join(source_path, 'qma.yml')
+    def template_config_file
+      source_path = File.expand_path('../../../config', __FILE__)
+      File.join(source_path, 'qma.yml')
+    end
+
+    def load_to_env
+      ENV['QMA_KEY'] = key
+      ENV['QMA_EXTERNAL_HOST'] = external_host
+      ENV['QMA_INTRANET_HOST'] = intranet_host
+    end
+
+    def update_host(type, host)
+      @data['host'] = {} unless @data.key?('host')
+      @data['host'][type.to_s] = host
+    end
+
+    def old_data?
+      @data.key?('development') || @data.key?('production')
+    end
+
+    def migratate_old_data
+      config = Config.new(template_config_file)
+
+      if external_host = @data.try(:[], 'production').try(:[], 'host')
+        config.external_host = external_host
+      else
+        config.external_host = nil
       end
 
-      def load_to_env
-        ENV['QMA_KEY'] = key
-        ENV['QMA_EXTERNAL_HOST'] = external_host
-        ENV['QMA_INTRANET_HOST'] = intranet_host
+      if intranet_host = @data.try(:[], 'development').try(:[], 'host')
+        config.intranet_host = intranet_host
+      else
+        config.intranet_host = nil
       end
 
-      def update_host(type, host)
-        @data['host'] = {} unless @data.has_key?('host')
-        @data['host'][type.to_s] = host
+      File.open(@path, 'w') do |f|
+        f.write config.data.to_yaml
       end
 
-
-      def old_data?
-        @data.has_key?('development') || @data.has_key?('production')
-      end
-
-      def migratate_old_data
-        config = Config.new(template_config_file)
-
-        if external_host = @data.try(:[], 'production').try(:[], 'host')
-          config.external_host = external_host
-        else
-          config.external_host = nil
-        end
-
-        if intranet_host = @data.try(:[], 'development').try(:[], 'host')
-          config.intranet_host = intranet_host
-        else
-          config.intranet_host = nil
-        end
-
-        File.open(@path, 'w') do |f|
-          f.write config.data.to_yaml
-        end
-
-        @data = config.data
-      end
-
-  end #/Config
-end #/QMA
+      @data = config.data
+    end
+  end # /Config
+end # /QMA
